+++ apm-agent-core/src/main/java/co/elastic/apm/agent/configuration/CoreConfiguration.java
@@ -253,6 +253,8 @@ public class CoreConfiguration extends ConfigurationOptionProvider {
             "\n" +
             "A few examples:\n" +
             "\n" +
+            " - `org.example.*` (omitting the method is possible since 1.4.0)\n" +
+            " - `org.example.*#*` (before 1.4.0, you need to specify a method matcher)\n" +
             " - `org.example.MyClass#myMethod`\n" +
             " - `org.example.MyClass#myMethod()`\n" +
             " - `org.example.MyClass#myMethod(java.lang.String)`\n" +
@@ -263,7 +265,9 @@ public class CoreConfiguration extends ConfigurationOptionProvider {
             "\n" +
             "NOTE: Only use wildcards if necessary.\n" +
             "The more methods you match to more overhead will be caused by the agent.\n" +
-            "Also note that there is a maximum amount of spans per transaction (see <<config-transaction-max-spans, `transaction_max_spans`>>).")
+            "Also note that there is a maximum amount of spans per transaction (see <<config-transaction-max-spans, `transaction_max_spans`>>).\n" +
+            "\n" +
+            "Since 1.3.0")
         .buildWithDefault(Collections.<MethodMatcher>emptyList());
 
     public boolean isActive() {