+++ apm-agent-core/src/test/java/co/elastic/apm/agent/configuration/AgentArgumentsConfigurationSourceTest.java
@@ -1,0 +1,38 @@
+/*-
+ * #%L
+ * Elastic APM Java agent
+ * %%
+ * Copyright (C) 2018 - 2019 Elastic and contributors
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package co.elastic.apm.agent.configuration;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class AgentArgumentsConfigurationSourceTest {
+
+    @Test
+    void testParse() {
+        assertThat(AgentArgumentsConfigurationSource.parse("foo=bar").getValue("foo")).isEqualTo("bar");
+        assertThat(AgentArgumentsConfigurationSource.parse("foo = bar ; baz =qux, quux ").getConfig())
+            .containsEntry("foo", "bar")
+            .containsEntry("baz", "qux, quux");
+        assertThatThrownBy(() -> AgentArgumentsConfigurationSource.parse("foo")).isInstanceOf(IllegalArgumentException.class);
+
+    }
+}