+++ apm-agent-core/src/main/java/co/elastic/apm/agent/bci/ElasticApmAgent.java
@@ -300,7 +300,7 @@ public class ElasticApmAgent {
             .ignore(any(), isReflectionClassLoader())
             .or(any(), classLoaderWithName("org.codehaus.groovy.runtime.callsite.CallSiteClassLoader"))
             .or(nameStartsWith("java.").and(not(nameStartsWith("java.util.concurrent."))))
-            .or(nameStartsWith("com.sun."))
+            .or(nameStartsWith("com.sun.").and(not(nameStartsWith("com.sun.faces."))))
             .or(nameStartsWith("sun"))
             .or(nameStartsWith("org.aspectj."))
             .or(nameStartsWith("org.groovy."))