+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-# next
+# next (1.4.0)
 
 ## Features
+ * The `trace_methods` configuration now allows to omit the method matcher.
+   Example: `com.example.*` traces all classes and methods within the `com.example` package and sub-packages.
 
 ## Bug Fixes