+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next (1.4.0)
 
 ## Features
+ * Added support for sync calls of OkHttp client
  * The `trace_methods` configuration now allows to omit the method matcher.
    Example: `com.example.*` traces all classes and methods within the `com.example` package and sub-packages.